added endpoints and integration testing. Many bugs still need to be resolved.

+++ src/main/java/com/jere/forum/repository/IRoleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IRoleRepository extends JpaRepository<RoleEntity, Long> {
+public interface IRoleRepository extends JpaRepository<RoleEntity, Integer> {
 
   RoleEntity findByName(String name);
