+++ src/main/java/com/jere/forum/repository/IUserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface IUserRepository extends JpaRepository<UserEntity, Long> {
+public interface IUserRepository extends JpaRepository<UserEntity, Integer> {
 
   UserEntity findByEmail(String email);
